+++ skycfg.go
@@ -22,9 +22,7 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"io/fs"
 	"os"
-	"path"
 	"path/filepath"
 	"strings"
 	"sync"
@@ -50,71 +48,6 @@ const (
 	contextKey   = "context"   // has type context.Context
 	logOutputKey = "logoutput" // has type io.Writer
 )
-
-// A FileReader controls how load() calls resolve and read other modules.
-type FileReader interface {
-	// Resolve parses the "name" part of load("name", "symbol") to a path. This
-	// is not required to correspond to a true path on the filesystem, but should
-	// be "absolute" within the semantics of this FileReader.
-	//
-	// fromPath will be empty when loading the root module passed to Load().
-	Resolve(ctx context.Context, name, fromPath string) (path string, err error)
-
-	// ReadFile reads the content of the file at the given path, which was
-	// returned from Resolve().
-	ReadFile(ctx context.Context, path string) ([]byte, error)
-}
-
-type localFileReader struct {
-	root string
-}
-
-// LocalFileReader returns a [FileReader] that resolves and loads files from
-// within a given filesystem directory.
-// LocalFileReader expects paths in load() to always use '/' as the separator,
-// regardless of the operating system's native path separator.
-func LocalFileReader(root string) FileReader {
-	if root == "" {
-		panic("LocalFileReader: empty root path")
-	}
-	return &localFileReader{root}
-}
-
-func (r *localFileReader) Resolve(ctx context.Context, name, fromPath string) (string, error) {
-	if fromPath == "" {
-		return name, nil
-	}
-	if filepath.Separator != '/' && strings.ContainsRune(name, filepath.Separator) {
-		return "", fmt.Errorf("load(%q): invalid character in module name", name)
-	}
-	resolved := filepath.Join(r.root, filepath.FromSlash(path.Clean("/"+name)))
-	return resolved, nil
-}
-
-func (r *localFileReader) ReadFile(ctx context.Context, path string) ([]byte, error) {
-	return os.ReadFile(path)
-}
-
-type fsFileReader struct {
-	fsys fs.FS
-}
-
-// FSFileReader returns a [FileReader] that loads files from the given [fs.FS].
-// Path resolution on a FSFileReader is just [path.Clean].
-func FSFileReader(fsys fs.FS) FileReader {
-	if fsys == nil {
-		panic("FSFileReader: nil fsys")
-	}
-	return &fsFileReader{fsys: fsys}
-}
-
-func (r *fsFileReader) Resolve(ctx context.Context, name, fromPath string) (string, error) {
-	return path.Clean(name), nil
-}
-
-func (r *fsFileReader) ReadFile(ctx context.Context, path string) ([]byte, error) {
-	return fs.ReadFile(r.fsys, path)
-}
 
 // NewProtoMessage returns a Starlark value representing the given Protobuf
 // message. It can be returned back to a proto.Message() via AsProtoMessage().
